Move tabulate to utils.misc and test it
