+++ src/pip/_internal/commands/list.py
@@ -7,7 +7,6 @@ import json
 import logging
 
 from pip._vendor import six
-from pip._vendor.six.moves import map, zip_longest
 
 from pip._internal.cli import cmdoptions
 from pip._internal.cli.req_command import IndexGroupCommand
@@ -18,13 +17,10 @@ from pip._internal.self_outdated_check import make_link_collector
 from pip._internal.utils.misc import (
     dist_is_editable,
     get_installed_distributions,
+    tabulate,
     write_output,
 )
 from pip._internal.utils.packaging import get_installer
-from pip._internal.utils.typing import MYPY_CHECK_RUNNING
-
-if MYPY_CHECK_RUNNING:
-    from typing import Any, Iterable, List, Tuple
 
 logger = logging.getLogger(__name__)
 
@@ -243,21 +239,6 @@ class ListCommand(IndexGroupCommand):
 
         for val in pkg_strings:
             write_output(val)
-
-
-def tabulate(rows):
-    # type: (Iterable[Iterable[Any]]) -> Tuple[List[str], List[int]]
-    """Return a list of formatted rows and a list of column sizes.
-
-    For example::
-
-    >>> tabulate([['foobar', 2000], [0xdeadbeef]])
-    (['foobar     2000', '3735928559'], [10, 4])
-    """
-    rows = [tuple(map(str, row)) for row in rows]
-    sizes = [max(map(len, col)) for col in zip_longest(*rows, fillvalue='')]
-    table = [" ".join(map(str.ljust, row, sizes)) for row in rows]
-    return table, sizes
 
 
 def format_for_columns(pkgs, options):
